+++ apps/docs/stories/theme.tsx
@@ -1,6 +1,0 @@
-const theme = `
-  body {  
-  }
-`;
-
-export default theme;