+++ apps/docs/stories/utils/decorators.ts
@@ -1,4 +1,4 @@
-import { API_PATHS, mockAllServices } from "./mockAllServices";
+import { setUpMocks } from './mockAllServices';
 
 type Props = { name: string; value: any }[];
 
@@ -22,18 +22,19 @@ const applyArgsToStoryComponent = (storyComponent: any, props: Props) => {
   });
 
   return component;
-}
+};
 
 const generateStyleBlock = (styleArg: any) => {
   const styleBlock = document.createElement('style');
   const styleArgKeys = Object.keys(styleArg);
 
   styleBlock.innerHTML = styleArgKeys.map((styleArgKey) => {
-    const selector = styleArgKey;
-    const cssProperties = styleArg[styleArgKey];
+      const selector = styleArgKey;
+      const cssProperties = styleArg[styleArgKey];
     const cssRules = Object.keys(cssProperties).map((cssProperty) => {
-      return `${cssProperty}: ${cssProperties[cssProperty]};`;
-    }).join('');
+          return `${cssProperty}: ${cssProperties[cssProperty]};`;
+        })
+        .join('');
 
     return (`
       ${selector} {
@@ -49,16 +50,7 @@ export const customStoryDecorator = (storyComponent: any, storyContext: any) => 
   const fragment = new DocumentFragment();
   const { props, styleArg } = getPropsAndStyles(storyContext);
 
-  const isMocksEnabled = __VITE_STORYBOOK_MOCKS_ENABLED__ === 'true';
-  const isChromaticBuild = __VITE_STORYBOOK_CHROMATIC_BUILD__ === 'true';
-
-  if (isMocksEnabled) {
-    // Use mock data for GrossPaymentChart only in Chromatic builds for consistent screenshots.
-    // For regular Storybook, use proxyApi to view dynamic data, especially to see dates from the past 30 days.
-    mockAllServices({
-      bypass: isChromaticBuild ? [] : [API_PATHS.GROSS_VOLUME],
-    });
-  }
+  setUpMocks();
 
   const component = applyArgsToStoryComponent(storyComponent, props);
 
@@ -69,4 +61,20 @@ export const customStoryDecorator = (storyComponent: any, storyContext: any) => 
 
   fragment.appendChild(component);
   return fragment;
+};
+
+// args is an object with key value pairs
+interface Args {
+  [key: string]: string;
+}
+export const getAttributesString = (args: Args) => {
+  // forEeach key in args return the attribute string only if it has a value
+
+  return Object.keys(args)
+    .map((key) => {
+      if (args[key]) {
+        return `${key}="${args[key]}"`;
+      }
+    })
+    .join(' ');
 };