+++ packages/webcomponents/src/components/business-forms/owner-form/owner-form.tsx
@@ -8,12 +8,13 @@ import { identitySchema } from '../schemas/business-identity-schema';
 import { config } from '../../../../config';
 import { LoadingSpinner } from '../../form/utils';
 import { deconstructDate } from '../utils/helpers';
-import { 
-  OwnerFormSubmitEvent, 
-  OwnerFormServerErrorEvent, 
-  OwnerFormServerErrors, 
-  OwnerFormClickEvent, 
-  OwnerFormClickActions } 
+import {
+  OwnerFormSubmitEvent,
+  OwnerFormServerErrorEvent,
+  OwnerFormServerErrors,
+  OwnerFormClickEvent,
+  OwnerFormClickActions
+}
   from '../utils/business-form-types';
 
 @Component({
@@ -42,6 +43,10 @@ export class BusinessOwnerForm {
 
   get identityEndpoint() {
     return this.ownerId ? `entities/identity/${this.ownerId}` : 'entities/identity';
+  }
+
+  get identificationNumberLabel() {
+    return this.owner.ssn_last4 ? 'Update SSN (optional)' : 'SSN';
   }
 
   get formTitle() {
@@ -88,6 +93,7 @@ export class BusinessOwnerForm {
       this.serverError.emit({ data: error, message: OwnerFormServerErrors.fetchData });
     } finally {
       this.isLoading = false;
+      console.log(this.formController.getInitialValues());
     }
   }
 
@@ -99,7 +105,7 @@ export class BusinessOwnerForm {
         const response = await this.api.patch(this.identityEndpoint, JSON.stringify(payload));
         return this.handleResponse(response);
       } else {
-        const payload = { ...parseIdentityInfo(this.formController.values.getValue()), business_id: this.businessId};
+        const payload = { ...parseIdentityInfo(this.formController.values.getValue()), business_id: this.businessId };
         const response = await this.api.post(this.identityEndpoint, JSON.stringify(payload));
         this.ownerId = response.data.id;
         return this.handleResponse(response);
@@ -112,7 +118,7 @@ export class BusinessOwnerForm {
   }
 
   handleResponse(response: any) {
-    this.submitted.emit({ data: response, metadata: { ownerId: this.ownerId }});
+    this.submitted.emit({ data: response, metadata: { ownerId: this.ownerId } });
     let errorMessage = this.ownerId ? OwnerFormServerErrors.patchData : OwnerFormServerErrors.postData;
     if (response.error) {
       this.serverError.emit({ data: response.error, message: errorMessage });
@@ -128,7 +134,7 @@ export class BusinessOwnerForm {
     if (!this.authToken) console.error(missingAuthTokenMessage);
 
     this.formController = new FormController(identitySchema('owner', this.allowOptionalFields));
-    this.api = Api(this.authToken, config.proxyApiOrigin);
+    this.api = Api({ authToken: this.authToken, apiOrigin: config.proxyApiOrigin });
     this.fetchData();
   }
 
@@ -269,7 +275,7 @@ export class BusinessOwnerForm {
               <div class="col-12 col-md-8">
                 <form-control-number-masked
                   name="identification_number"
-                  label="SSN"
+                  label={this.identificationNumberLabel}
                   defaultValue={ownerDefaultValue?.identification_number}
                   error={this.errors.identification_number}
                   inputHandler={this.inputHandler}