+++ packages/webcomponents/src/components/checkout/checkout-core.tsx
@@ -3,6 +3,7 @@ import { extractComputedFontsToLoad, formatCurrency } from '../../utils/utils';
 import { config } from '../../../config';
 import { PaymentMethodPayload } from './payment-method-payload';
 import { Checkout, ICheckout, ICheckoutCompleteResponse } from '../../api/Checkout';
+import { ComponentError, ComponentErrorCodes, ComponentErrorSeverity } from '../../api/ComponentError';
 
 @Component({
   tag: 'justifi-checkout-core',
@@ -19,15 +20,19 @@ export class CheckoutCore {
   @Prop() complete: Function;
   @Prop() checkoutId: string;
 
+  @Prop() disableCreditCard?: boolean;
+  @Prop() disableBankAccount?: boolean;
+  @Prop() disableBnpl?: boolean;
+  @Prop() disablePaymentMethodGroup?: boolean;
+
   @State() hasLoadedFonts: boolean = false;
   @State() isLoading: boolean = false;
   @State() checkout: ICheckout;
   @State() serverError: boolean = false;
-  @State() errorMessage: string = '';
   @State() creatingNewPaymentMethod: boolean = false;
-  @State() selectedPaymentMethodToken: string;
 
-  @Event() submitted: EventEmitter<ICheckoutCompleteResponse>;
+  @Event({ eventName: 'submitted' }) submitted: EventEmitter<ICheckoutCompleteResponse>;
+  @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentError>;
 
   private paymentMethodOptionsRef?: HTMLJustifiPaymentMethodOptionsElement;
 
@@ -46,16 +51,19 @@ export class CheckoutCore {
 
   fetchData(): void {
     this.isLoading = true;
-
     this.getCheckout({
       onSuccess: ({ checkout }) => {
         this.checkout = new Checkout(checkout);
         this.isLoading = false;
       },
-      onError: (errorMessage) => {
-        this.errorMessage = errorMessage;
+      onError: ({ error, code, severity }) => {
         this.isLoading = false;
-      },
+        this.errorEvent.emit({
+          errorCode: code,
+          message: error,
+          severity,
+        });
+      }
     });
   };
 
@@ -83,25 +91,43 @@ export class CheckoutCore {
 
     const payload: PaymentMethodPayload = await this.paymentMethodOptionsRef.resolvePaymentMethod();
 
-    if (payload.token) {
+    if (!payload) {
+      this.isLoading = false;
+    }
+    else if (payload.error) {
+      this.onError({
+        code: (payload.error.code as ComponentErrorCodes),
+        error: payload.error.message,
+        severity: ComponentErrorSeverity.ERROR,
+      });
+    }
+    else if (payload.token) {
       this.complete({
         payment: { payment_mode: 'ecom', payment_token: payload.token },
         onSuccess: this.onSubmitted,
-        onError: this.onSubmitted,
+        onError: this.onError,
       })
-    } else if (payload.bnpl?.status === 'success') {
+    }
+    else if (payload.bnpl?.status === 'success') {
       this.complete({
         payment: { payment_mode: 'bnpl' },
         onSuccess: this.onSubmitted,
-        onError: this.onSubmitted,
+        onError: this.onError,
       })
-    } else {
-      this.isLoading = false;
     }
   }
 
-  onSubmitted = (data) => {
+  onSubmitted = (data: ICheckoutCompleteResponse) => {
     this.submitted.emit(data);
+    this.isLoading = false;
+  };
+
+  onError = ({ error, code, severity }: { error: string, code: ComponentErrorCodes, severity: ComponentErrorSeverity }) => {
+    this.errorEvent.emit({
+      errorCode: code,
+      message: error,
+      severity,
+    });
     this.isLoading = false;
   };
 
@@ -135,8 +161,10 @@ export class CheckoutCore {
             <div class="d-flex flex-column">
               <justifi-payment-method-options
                 ref={(el) => (this.paymentMethodOptionsRef = el)}
-                show-card={this.checkout?.payment_settings?.credit_card_payments || true}
-                show-ach={this.checkout?.payment_settings?.ach_payments || true}
+                show-card={!this.disableCreditCard}
+                show-ach={!this.disableBankAccount}
+                show-bnpl={!this.disableBnpl}
+                show-saved-payment-methods={!this.disablePaymentMethodGroup}
                 bnpl={this.checkout?.bnpl}
                 client-id={this.checkout?.payment_client_id}
                 account-id={this.checkout?.account_id}