+++ packages/webcomponents/src/components/checkout/checkout.tsx
@@ -1,18 +1,25 @@
-import { Component, Prop, h, State, Watch } from '@stencil/core';
+import { Component, Prop, h, State, Watch, Event, EventEmitter } from '@stencil/core';
 import { makeGetCheckout, makeCheckoutComplete } from './checkout-actions';
 import { CheckoutService } from '../../api/services/checkout.service';
+import { ComponentError, ComponentErrorCodes, ComponentErrorSeverity } from '../../api/ComponentError';
 
 @Component({
-  tag: 'justifi-checkout',
-  shadow: true,
+  tag: 'justifi-checkout'
 })
 export class Checkout {
   @Prop() iframeOrigin?: string;
   @Prop() authToken: string;
   @Prop() checkoutId: string;
+  @Prop() disableCreditCard?: boolean;
+  @Prop() disableBankAccount?: boolean;
+  @Prop() disableBnpl?: boolean;
+  @Prop() disablePaymentMethodGroup?: boolean;
+
   @State() getCheckout: Function;
   @State() complete: Function;
   @State() errorMessage: string = '';
+
+  @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentError>;
 
   componentWillLoad() {
     this.initializeGetCheckout();
@@ -31,7 +38,11 @@ export class Checkout {
         service: new CheckoutService()
       });
     } else {
-      this.errorMessage = 'auth-token and checkout-id are required';
+      this.errorEvent.emit({
+        message: 'auth-token and checkout-id are required',
+        errorCode: ComponentErrorCodes.MISSING_PROPS,
+        severity: ComponentErrorSeverity.ERROR,
+      });
     }
   }
 
@@ -46,6 +57,10 @@ export class Checkout {
       <justifi-checkout-core
         getCheckout={this.getCheckout}
         complete={this.complete}
+        disableCreditCard={this.disableCreditCard}
+        disableBankAccount={this.disableBankAccount}
+        disableBnpl={this.disableBnpl}
+        disablePaymentMethodGroup={this.disablePaymentMethodGroup}
       ></justifi-checkout-core>
     );
   }