+++ packages/webcomponents/src/components/checkout/payment-method-options.tsx
@@ -13,19 +13,21 @@ import { IBnpl } from '../../api';
 export class PaymentMethodOptions {
   @Prop() showCard: boolean;
   @Prop() showAch: boolean;
+  @Prop() showBnpl: boolean;
+  @Prop() showSavedPaymentMethods: boolean;
   @Prop() bnpl: IBnpl;
   @Prop() clientId: string;
   @Prop() accountId: string;
   @Prop({ mutable: true }) iframeOrigin?: string = config.iframeOrigin;
   @Prop() savedPaymentMethods: any[] = [];
-  @Prop() selectedPaymentMethodId: string;
   @Prop() paymentAmount: string;
 
+  @State() selectedPaymentMethodId: string;
   @State() paymentMethodOptions: PaymentMethodOption[] = [];
 
   @Event({ bubbles: true }) toggleCreatingNewPaymentMethod: EventEmitter;
 
-  private selectedPaymentMethodOptionRef?: HTMLJustifiNewPaymentMethodElement | HTMLJustifiSavedPaymentMethodElement | HTMLJustifiSezzelPaymentMethodElement;
+  private selectedPaymentMethodOptionRef?: HTMLJustifiNewPaymentMethodElement | HTMLJustifiSavedPaymentMethodElement | HTMLJustifiSezzlePaymentMethodElement;
 
   @Watch('savedPaymentMethods')
   paymentMethodsChanged() {
@@ -36,8 +38,8 @@ export class PaymentMethodOptions {
     if (this.showAch) {
       this.paymentMethodOptions.push(new PaymentMethodOption({ id: PaymentMethodTypes.bankAccount }));
     }
-    if (this.bnpl?.provider === 'sezzle') {
-      this.paymentMethodOptions.push(new PaymentMethodOption({ id: PaymentMethodTypes.sezzel }));
+    if (this.showBnpl && this.bnpl?.provider === 'sezzle') {
+      this.paymentMethodOptions.push(new PaymentMethodOption({ id: PaymentMethodTypes.sezzle }));
     }
     this.selectedPaymentMethodId = this.paymentMethodOptions[0].id;
   }
@@ -59,7 +61,7 @@ export class PaymentMethodOptions {
           const newCard = paymentMethodOption.id === PaymentMethodTypes.card;
           const newBankAccount = paymentMethodOption.id === PaymentMethodTypes.bankAccount;
           const isSelected = this.selectedPaymentMethodId === paymentMethodOption.id;
-          const sezzel = paymentMethodOption.id === PaymentMethodTypes.sezzel;
+          const sezzle = paymentMethodOption.id === PaymentMethodTypes.sezzle;
           if (newCard || newBankAccount) {
             return (
               <justifi-new-payment-method
@@ -74,9 +76,9 @@ export class PaymentMethodOptions {
                 }}
               />
             );
-          } else if (sezzel) {
+          } else if (sezzle) {
             return (
-              <justifi-sezzel-payment-method
+              <justifi-sezzle-payment-method
                 paymentMethodOption={paymentMethodOption}
                 is-selected={isSelected}
                 paymentAmount={this.paymentAmount}
@@ -86,9 +88,9 @@ export class PaymentMethodOptions {
                     this.selectedPaymentMethodOptionRef = el;
                   }
                 }}>
-              </justifi-sezzel-payment-method>);
+              </justifi-sezzle-payment-method>);
           }
-          else {
+          else if (this.showSavedPaymentMethods) {
             return (
               <justifi-saved-payment-method
                 paymentMethodOption={paymentMethodOption}